+++ NavigationReactNative/sample/iosTabs/Scene.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {onNavigate} from './NavigationMotion.js';
+import {AppRegistry, NativeEventEmitter, NativeModules} from 'react-native';
 import {StateNavigator, StateContext} from 'navigation';
 import {NavigationContext} from 'navigation-react';
 
@@ -7,9 +7,6 @@ class Scene extends React.Component {
     constructor(props) {
         super(props);
         this.state = {navigationEvent: null};
-    }
-    static defaultProps = {
-        crumb: 0
     }
     static getDerivedStateFromProps(props, {navigationEvent: prevNavigationEvent}) {
         var {crumb, navigationEvent} = props;
@@ -49,10 +46,20 @@ class Scene extends React.Component {
         return stateContext;
     }
     componentDidMount() {
-        if (!this.props.crumb) {
-            this.subscription = onNavigate(({crumb, tab}) => {
+        var {crumb, tab, navigationEvent: {stateNavigator}} = this.props;
+        if (!crumb) {
+            var {NavigationMotion} = NativeModules;
+            this.handleNavigation = (_oldState, _state, _data, _asyncData, stateContext) => {
+                var {crumbs, title, history} = stateContext;
+                if (!history) {
+                    var titles = crumbs.map(({title}) => title).concat(title);
+                    NavigationMotion.render(crumbs.length, tab, titles, AppRegistry.getAppKeys()[0]);
+                }
+            }
+            stateNavigator.onNavigate(this.handleNavigation); 
+            var emitter = new NativeEventEmitter(NavigationMotion)
+            this.subscription = emitter.addListener('Navigate', ({crumb, tab}) => {
                 if (this.props.tab === tab) {
-                    var {stateNavigator} = this.props.navigationEvent;
                     var distance = stateNavigator.stateContext.crumbs.length - crumb;
                     if (distance > 0) {
                         var url = stateNavigator.getNavigationBackLink(distance);
@@ -66,17 +73,18 @@ class Scene extends React.Component {
         return state.navigationEvent === props.navigationEvent;
     }
     componentWillUnmount() {
-        if (!this.props.crumb)
+        var {crumb, navigationEvent: {stateNavigator}} = this.props;
+        if (!crumb) {
+            stateNavigator.offNavigate(this.handleNavigation); 
             this.subscription.remove();
+        }
     }
     render() {
-        var {crumb, navigationEvent, tab, ...props}  = this.props;
         var {navigationEvent} = this.state;
-        if (!navigationEvent) return null;
         var {state, data} = navigationEvent.stateNavigator.stateContext;
         return (
             <NavigationContext.Provider value={navigationEvent}>
-                {state.renderScene(data, props)}
+                {state.renderScene(data)}
             </NavigationContext.Provider>
         );
     }
